Testing scClassify model

Testing scClassify model

predict_scClassify(exprsMat_test, trainRes, cellTypes_test = NULL,
  k = 10, prob_threshold = 0.7, cor_threshold_static = 0.5,
  cor_threshold_high = 0.7, features = "limma", algorithm = "WKNN",
  similarity = "pearson", cutoff_method = c("dynamic", "static"),
  parallel = FALSE, ncores = 1, verbose = FALSE)

Arguments

exprsMat_test
A list or a matrix indicates the expression matrices of the query datasets

trainRes
A `scClassifyTrainModel` or a `list` indicates scClassify trained model

cellTypes_test
A list or a vector indicates cell types of the qurey datasets (Optional).

k
An integer indicates the number of neighbour

prob_threshold
A numeric indicates the probability threshold for KNN/WKNN/DWKNN.

cor_threshold_static
A numeric indicates the static correlation threshold.

cor_threshold_high
A numeric indicates the highest correlation threshold

features
A vector indicates the gene selection method, set as "limma" by default. This should be one or more of "limma", "DV", "DD", "chisq", "BI".

algorithm
A vector indicates the KNN method that are used, set as "WKNN" by default. This should be one or more of "WKNN", "KNN", "DWKNN".

similarity
A vector indicates the similarity measure that are used, set as "pearson" by default. This should be one or more of "pearson", "spearman", "cosine", "jaccard", "kendall", "binomial", "weighted_rank","manhattan"

cutoff_method
A vector indicates the method to cutoff the correlation distribution. Set as "dynamic" by default.

parallel
A logical input indicates whether running in paralllel or not

ncores
An integer indicates the number of cores that are used

verbose
A logical input indicates whether the intermediate steps will be printed

Value

list of results